corrected lr int

+++ examples/minChatGPT/reward_modeling.py
@@ -36,7 +36,7 @@ class ScriptArguments:
     per_device_train_batch_size: Optional[int] = field(default=1)
     per_device_eval_batch_size: Optional[int] = field(default=2)
     gradient_accumulation_steps: Optional[int] = field(default=4)
-    learning_rate: Optional[int] = field(default=2e-5)
+    learning_rate: Optional[float] = field(default=2e-5)
     weight_decay: Optional[int] = field(default=0.001)
     model_name: Optional[str] = field(
         default="gpt2",
